+++ apps/demo/urls.py
@@ -6,4 +6,5 @@ app_name = "demo"
 
 urlpatterns = [
     path("", views.demo_center_view, name="center"),
+    path("report/", views.demo_report_view, name="report"),
 ]